Refactor k6 test structure
