+++ package.json
@@ -1,15 +1,12 @@
 {
   "name": "test-big-chunk-front-k9",
   "version": "1.0.0",
-  "description": "test big chunk of actions in front using k9",
+  "description": "test big chunk of actions using k6",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "load-test": "k6 run tests/login.k6.js"
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "k9": "^0.0.2"
-  }
+  "license": "ISC"
 }
